TASK_gpt-4o_Hallucination/requirements.txt
openai
pandas
datasets
langchain
faiss-cpu
scikit-learn
transformers
python-dotenv
lettucedetect
protobuf
sentencepiece
matplotlib